k256: impl `signature` traits for `schnorr::*Key`
The traits provide standard, high-level interfaces for
signing/verification.

They also take care of computing a (SHA-256) digest from an input
message.
